// packages/app/api/generated/models/CreateSubject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateSubject
 */
export interface CreateSubject {
    /**
     * 
     * @type {string}
     * @memberof CreateSubject
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CreateSubject
     */
    costPerHour?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateSubject
     */
    availability?: string;
}

/**
 * Check if a given object implements the CreateSubject interface.
 */
export function instanceOfCreateSubject(value: object): value is CreateSubject {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateSubjectFromJSON(json: any): CreateSubject {
    return CreateSubjectFromJSONTyped(json, false);
}

export function CreateSubjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateSubject {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'costPerHour': json['costPerHour'] == null ? undefined : json['costPerHour'],
        'availability': json['availability'] == null ? undefined : json['availability'],
    };
}

export function CreateSubjectToJSON(json: any): CreateSubject {
    return CreateSubjectToJSONTyped(json, false);
}

export function CreateSubjectToJSONTyped(value?: CreateSubject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'costPerHour': value['costPerHour'],
        'availability': value['availability'],
    };
}