// packages/app/api/generated/models/Subject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Subject
 */
export interface Subject {
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof Subject
     */
    costPerHour?: number;
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    availability?: string;
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    tutorId?: string;
}

/**
 * Check if a given object implements the Subject interface.
 */
export function instanceOfSubject(value: object): value is Subject {
    return true;
}

export function SubjectFromJSON(json: any): Subject {
    return SubjectFromJSONTyped(json, false);
}

export function SubjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): Subject {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'costPerHour': json['costPerHour'] == null ? undefined : json['costPerHour'],
        'availability': json['availability'] == null ? undefined : json['availability'],
        'tutorId': json['tutorId'] == null ? undefined : json['tutorId'],
    };
}

export function SubjectToJSON(json: any): Subject {
    return SubjectToJSONTyped(json, false);
}

export function SubjectToJSONTyped(value?: Subject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'costPerHour': value['costPerHour'],
        'availability': value['availability'],
        'tutorId': value['tutorId'],
    };
}