// packages/app/api/generated/models/UpdateSubject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateSubject
 */
export interface UpdateSubject {
    /**
     * 
     * @type {string}
     * @memberof UpdateSubject
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateSubject
     */
    costPerHour?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateSubject
     */
    availability?: string;
}

/**
 * Check if a given object implements the UpdateSubject interface.
 */
export function instanceOfUpdateSubject(value: object): value is UpdateSubject {
    return true;
}

export function UpdateSubjectFromJSON(json: any): UpdateSubject {
    return UpdateSubjectFromJSONTyped(json, false);
}

export function UpdateSubjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateSubject {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'costPerHour': json['costPerHour'] == null ? undefined : json['costPerHour'],
        'availability': json['availability'] == null ? undefined : json['availability'],
    };
}

export function UpdateSubjectToJSON(json: any): UpdateSubject {
    return UpdateSubjectToJSONTyped(json, false);
}

export function UpdateSubjectToJSONTyped(value?: UpdateSubject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'costPerHour': value['costPerHour'],
        'availability': value['availability'],
    };
}